swap around the row applier buffering in order to fix a memory-hungry case.
previously in the case where the 'from' end had deleted a large set of rows, the 'to' end would sometimes load those rows into memory all at once only to discover that they needed to be removed.  this could require a potentially unbounded amount of memory, which the design seeks to avoid.

to fix this, buffer the source rows instead of the local rows, and process the local rows in the same ranges without buffering the query results; if the source rows are big enough, process them in batches rather than buffering the whole lot at once.  doing it this way around fixes the above case and requires fewer special cases in the logic.