Added '--structure-only' flag
This will create the table structure without fetching any of the actual
row data.

The purpose of this is so that one can clone only the database
'skeleton', so that the tables can be populated manually.

This will allow 'partial-ks' to run successfully on a machine that
hasn't been kitchen sync'd before, as previously every table had to
exist so that a 'schema.rb' file could be generated.