Greet the SSH server with a NOP so database can be added to known hosts
Before this, ks would spin up multiple workers and would not allow any of them
to accept the host key fingerprint.

After this change, the host key prompt is asked before any syncing
actually happens. This ensures that the user always has a chance to
accept the host key fingerprint if necessary.

The SSH greeting will not prompt for anything if the host key
fingerprint is already known.