add support for decoding msgpack type 0xd9
Not sure why we didn't have this at the time, perhaps the ancient docs didn't have this type.

Although we can now ready it, we Won't start *writing* using this type until we do a breaking protocol change, because it would produce different hash results whenever we encounter a string between 33 and 255 characters long.