support postgresql's UUID type
Honestly I'm surprised no-one asked for this already!

There's multiple sensible things we could map it to on MySQL/MariaDB, and we don't yet have any mechanism to choose, so I had to just pick one option; I've chosen to map it to CHAR(36), on the assumption that this is the most likely to be immediately usable for people doing porting of applications that don't do anything in particular for UUIDs.