fix a test case that randomly freezes up on the mysql 8.0 build
This was showing up as one database connection blocked with state "Waiting for commit lock" in "SHOW KEYS FROM footbl" with the other database connection idle.  The actual KS 'to' code doesn't send the schema retrieval commands until the leader has un-held the snapshot, so do the same in the test case to resolve this.