when converting mysql column values, only tinyint(1) should be interpreted as boolean (closes #56)
start memoizing the mysql column value conversion decisions rather than figuring it out as we iterate over each individual value, since it's getting less trivial as time goes on.

also begin verifying that tinyint(1) values are actually "0" (false) or "1" (true) rather than silently converting any other value to false.

remove the special case that attempted to treat tinyint(1) columns with defaults other than "0" or "1" as integer columns; this didn't actually work, per issue #56, and it was too brittle - it would be silly for users to suddenly get broken data syncs when the default happens to change.

at some point we will probably need some kind of configurability for the assumption that tinyint(1) is best treated as boolean.  since this is the first user report of anyone trying even tinyint(2), we'll wait until we hear of an actual use case before we look into customizable mappings.