move where substitute primary key selection is done; extract these special-purpose methods to their own file; ignore differences in the choice when schema matching
moving this for two reasons.  firstly, since the fallback case's choice isn't really a fundamental related to the underlying database, it isn't really right to do it in the database adapter.

secondly, it was unclear that the choice at the 'to' end doesn't really matter because the 'to' end will always actually use the copy of the schema received from the 'from' end to run its queries.

this revealed a missed opportunity to keep the table (rather than recreating it) in the case where the chosen key would change (for example, because there was no suitable key at the 'to' end previously) - as the substitute key was not reconsidered after changing the key ist, match_table's final (from_table == to_table) comparison would return false and the table would then be recreated.  added a test to cover this case, and changed the implementation of Table#== to ignore the list when comparing tables that don't have explicit primary keys.