best-efforts support for tables with no primary key and no suitable substitute unique key (closes #58)
try to select a partial key if there's no usable primary key, and if possible use that as best we can (comparing hashes for groups of rows without finding individual rows that don't match).  selecting a key is itself slow, as we query the table to find the key with the best cardinality.

if we can't even find a suitable partial key, clear and reload the whole table, rather than erroring out.