to choose a partial key, use the database server's own column cardinality estimates instead of explicitly querying (#58)
this is significantly less accurate - it depends on the table having been ANALYZEd at some point and it will give somewhat bogus estimates data for multi-part keys, which are probably most keys found on tables with no real PK - but it will be much faster to start up for these tables with no real PK and no suitable substitute, where we're trying to pick a partial key to use instead.

this is important since we unfortunately run all the partial-key-selection code on *each* 'from' worker (because they are not in contact with one another), meaning the impact of running the slow key count queries could have been quite heavy.

the possibility of it guessing wrong is just another good reason why you should always have proper PKs!