fix inefficient extra call when encoding bytea values (closes #68)
we weren't chopping off the null terminator returned by PQescapeByteaConn, so we got unterminated quoted string SQL syntax errors if we didn't call PQescapeStringConn - but that was only working around the problem.

perhaps surprisingly, bytea values still seemed to be round-tripping correctly (according to our test cases such as "handles requesting and saving arbitrary binary values in BLOB fields").  I believe this is because a) the string retured by PQescapeByteaConn is always simply "\x" + the hex-encoded input string, and so never contains a single-quote character that gets quoted by PQescapeStringConn and b) the (now escaped to "\u0000" in order to make it into an SQL statement) null byte after the hex-encoded string was effectively a terminator since it wasn't part of the hex-encoded string.

so no harm except needless inefficiency.