workaround MySQL 8 swapping axis order on input of geometry in SRSs defined officially as lat-lon
Unfortunately the most commonly used SRS, 4326 which corresponds to the WGS84 datum, is so defined, so this is in practice a significant change.

Irrespective of input format/SRS, MySQL 8 always stores in lon-lat format, and so that's what we get back if we do a raw select (as we do); ST_AsBinary/ST_AsWKB would swap the order back again, but then we'd need to separately select the SRID as well in order to get all the data we need to pass across, which is awkward (and as per comments in previous commit, hard to combine with any anonymization rules the user might happen to enter).

And, that still wouldn't really be compatible; in particular, PostGIS always uses lon-lat format, for both input and output.  Earlier versions of MySQL and MariaDB are in some ways effectively agnostic as they don't really use SRSs (according to docs even those versions that can store SRIDs ignored it and act as if everything is SRID 0 ie. a flat cartesian plane), but did have a few functions that assumed that the data was already lon-lat.

So, it's better for us to declare that our wire format always lon-lat and tell MySQL that in tests (when we go to insert the WKT data or select back WKB data).  (Conveniently, this means we don't actually need to do any conversion in the actual sync for any version of MySQL, since wire, INSERT, and SELECT format will all be the raw binary for which is lon-lat in all cases.)

It's not certain that this is always the right thing to do when syncing from mysql <8 (or mariadb) to mysql >8, but it is our best guess in that case and should be the most useful behavior for the other cases.  Note that similar potential for confusion arises when upgrading or replicating between such versions as well, so we are matching what they do - which seems in line with user expectations.  Note also that since SRIDs didn't actually do anything in previous versions, it's actually quite common for them to be unspecified, in which case axis order swapping will not have happened and it's moot.

See https://mysqlserverteam.com/spatial-reference-systems-in-mysql-8-0/ for details on the MySQL 8.0 changes, and http://docs.geotools.org/latest/userguide/library/referencing/order.html for the history of the lon-lat vs. lat-lon issue in geospatial systems.