implement schema support for spatial indexes
This is potentially sending us down the slippery slope towards trying to represent all the schema stuff that KS doesn't really have any business caring about.

Arguably, we were already one step down the start of this slope because we have always represented all the keys on the tables, even those that are not unique - and really unique keys are the only ones directly relevant to KS, because they are the only ones that need to be consistent in order for syncing to succeed (and the only ones that KS needs to do anything different for - clearing conflicting rows).

If we did not extend our key representation to know about spatial keys, we would potentially try to create useless normal keys on spatial columns, and now that we know about spatial column types it seems likely that users would expect us to know about spatial indexes too, and be able to correctly create tables containing them from scratch.

So slippery slope or not, this seems worth doing.