switch from MySQL's prefixed format to PostgreSQL's EWKB format
As previously mentioned in the comments, it's somewhat arbitrary.  I had a change of heart mainly because PostGIS is so much more widely used than MySQL's spatial type support, so it makes more sense to take it as the de-facto standard.

In addition, PostGIS's EWKB format is a superset of MySQL's WKB format (eg. PostGIS supports a third dimension); using the SRID-prefixed format meant we were sometimes going to be dealing in values that were neither PostGIS-style EWKB data nor legal MySQL data, which seems silly when one format is a superset of the other.