implement JSON column type
On PostgreSQL and MySQL 8+, this uses the native json type.
On MariaDB 10.2+, this uses LONGTEXT with a json_valid CHECK constraint (as MariaDB itself would use for the JSON type alias in later versions).
On MySQL 5.5 and MariaDB 5.5 through 10.1, this uses LONGTEXT with a column comment to indicate that it's for JSON content (mainly for the benefit of our test suite).