implement support for enumeration columns (closes #70)
For PostgreSQL, we haven't implemented enumerated type creation/maintainance, so the appropriate enumerated types must exist already.

I've chosen to keep serializing the values as strings rather than index integers, as that is how the native database client libraries actually return them to us.  Theoretically it would be more efficient if we were to index them, since then we'd need to hash fewer bytes, but it's nontrivial to implement for mysql/mariadb as the MYSQL_FIELD structures returned in the resultsets don't give the enumeration values, and there are no formal type IDs/names as for postgresql, so we'd need to start passing around the Column structures to the packers, which is quite a bit of work, and the increase in complexity doesn't seem worthwhile, especially given how uncommon enumeration columns are compared to other types which don't need all that stuff.