fix unnecessary recreation of tables with enums when syncing between mysql and postgresql
we don't have adequate test coverage here.  we currently test that tables are created/modified in a way that matches, but not that the schema matcher considers them equal afterwards.  differences between the two that are not amendable (such as whether enumeration types are named) need to be worked around by these steps, but that's hard to expose in tests.