fix setting a default and changing a column to be non-nullable at the same time
Previously on mysql we would try and do this:
	"ALTER TABLE `footbl` MODIFY `another_col` smallint NOT NULL DEFAULT 23, ALTER `another_col` SET DEFAULT 23"
which is clearly redundant but also returns an error:
	"Unknown column 'another_col' in 'footbl'"

PostgreSQL wasn't affected as has the SET/DROP NOT NULL syntax, so its first clause is different and the second clause is therefore still valid:
	"ALTER TABLE "footbl" ALTER "another_col" SET NOT NULL, ALTER "another_col" SET DEFAULT 23"

The fix for the MySQL case here is not very obvious.  The problem was that the MODIFY statement we run does actually completely redefine the column, but because it wasn't correctly updating default_value to show that that had already been changed, the next step in the matcher code thought it needed to change that default still.  Now that we show that that has already been done, it will simply run:
	"ALTER TABLE `footbl` MODIFY `another_col` smallint NOT NULL DEFAULT 23"

I've also extracted out ModifyColumnDefinitionClauses just to better explain what actually happens (ie. that changing the default at the same time as the nullability is implemented by modifying the column as a whole).