remove the warning for masked flags
Although a nice idea in a lot of ways, there's no particular reason why we should do this for things that are represented by flags and not do it for the things represented by other Column attributes.

We could output a similar warning for the things changed in convert_unsupported_database_schema, but many of those are promotions (eg. width increases preserving type) and some not (eg. replacing UUID type with a character field), and it's hard to say which changes are relevant enough to warn about.

Complaining on every run is probably not desirable for many users, so it doesn't work well enough at the moment to be worth keeping, and it's better to be simple and consistent.