portable support for fractional seconds in time/datetime/timestamp/timestamptz values (closes #74)
mostly this affects mysql/mariadb, as postgresql only/always supports microsecond precision in time and date-time types.  mysql has supported fractional seconds precision since 8.0 and mariadb since 5.3 (partial)/10.0 (full including the relevant functions), but they're still not the default; you can choose the precision (from 0 to 6 decimal places).

after this work, we will maintain the precision when syncing from mysql/mariadb to another instance of mysql/mariadb (provided the 'to' end is one of the versions above), and will use 6 dp. precision when syncing from postgresql to mysql/mariadb (again, for supported versions); this is effectively a behavior change, since previously we would use the traditional 0 dp. types when syncing from postgresql to mysql/mariadb.

to get the hashes of time/datetime values to match, we trim off trailing zeros (and if necessary the decimal point), like postgresql format.

getting the default/on update definitions to be accepted on every version of mysql & mariadb was a challenge but the test suite now passes for all of them.