fold sync_from_protocol.h back into sync_from.h
This was split out at the time where we moved the sync algorithm from being something that ran at both ends to something that ran only at the 'to' end.  Consequently, the version of protocol actually dictated whether and which sync algorithm was run, so it was implemented quite separately.

Support for that old version was long ago removed, so there's no reason to have half of the (now much simpler and more enduring) 'from' protocol code in a separate file.