unset NO_ZERO_DATE for the benefit of old mysql/mariadb versions
the old version in the centos build in the test suite revealed they are unable to run ADD COLUMN ... DATE NOT NULL when NO_ZERO_DATE is set, as they implicitly try to set the date fields to '0000-00-00'.  we do this to avoid needlessly rewriting the entire table an extra time (though this optimisation only has full effect on versions recent enough to have instant ADD COLUMN), and since we will overwrite the values during sync later, this should not result in any data issues afterwards.

we don't have any particular need for ERROR_FOR_DIVISION_BY_ZERO, but it was part of the previous value (TRADITIONAL), so retain it until we have a good reason not to.  STRICT_TRANS_TABLES and STRICT_ALL_TABLES are very much desirable to catch problems, and I consider PIPES_AS_CONCAT useful for portability (though I accept this should probably be controlled by an option somewhere or off by default, as I use it in my apps but not everyone does).