Send a list of supported types from protocol v8 on (#60)
The main goal here is to be able to add new types later at the 'from' end and not surprise clients that don't understand them.

Although currently sending a column with a new type unknown to the 'to' end wouldn't really behave any worse than the explicit UNKN type, after #60 the 'to' end will be able to utilize db_type_def to support UNKN types - but would not know what to use for a new (actually unknown) type string, which is worse.  This change will make it possible to avoid receiving them and thus ensure that adding a new type at the 'from' end won't degrade behavior at the 'to' end.

As a side bonus, this will disable the new types such as JSON and SPAT for v7 clients, so they'll continue to get the legacy error behavior.