Complete schema column-type system rewrite for protocol version 8
Types are now negotiated; the 'to' end supplies its list of supported types and the 'from' end uses this to downgrade from database-specific types to more generic types.  This replaces the use of column flags to indicate the database-specific variants of types, simplifying schema serialization and extensibility.

Types are still represented on the wire as strings for simplicity, but are mapped to integer codes in memory in order to speed up comparisons (we can now use switch statements instead of if/else string comparison statements).

I've backed out some of the unreleased type ideas such as using COMMENT fields on columns to mark UUID and JSON columns on versions of mysql where they are not supported, which I was starting to get cold feet on as it seemed a bit too hacky/KS-specific.

I've introduced 'mysql-specific' and 'postgresql-specific' types, separate to 'unknown'.  This will allow us to use separate mappings for unknown types in the future.