Replace the last remaining flag with an enum
This gives us a place to put things like SQL Server's rowversion (which is otherwise just an 8-byte binary field) if we add support for other databases in the future.

Ideally we wouldn't have had to have a special type for MySQL's own timestamp either, but it has a separate datetime type, and either can be used with the default and/or ON UPDATE behavior.