Use the 'replica' session_replication_role on PostgreSQL to disable foreign key constraints; stop trying 'SET CONSTRAINTS ALL DEFERRED'
This is a very long-standing TODO. We really want a direct equivalent to mysql's foreign_key_checks, but there's no such unprivileged setting available in PostgreSQL; this is the nearest equivalent, and like using ALTER TABLE ... DISABLE TRIGGER it requires superuser privileges, so we can't use it all the time.

'SET CONSTRAINTS ALL DEFERRED' wasn't really very effective because it only deferred until commit, and under --commit=often it wouldn't last long enough to help avoid conflicts partway through the table.

There are 6 types of constraints:
* CHECK constraints - which were not affected by this setting and we don't need to do anything special to avoid them
* NOT NULL constraints - which were not affected by this setting and we don't need to do anything special to avoid them
* PRIMARY KEY constraits - which are fine for us
* UNIQUE constraints - which are explicitly dealt with (by our code to 'clear' conflicting unique indexes values)
* EXCLUDE constraints - which are still not handled at this time, but in my testing don't work properly with deferral anyway (INITALLY DEFERRED never gets checked, and INITIALLY IMMEDIATE with SET CONSTRAINTS ... DEFERRED doesn't work)
* FOREIGN KEY constraints - which are now handled by replica session_replication_role.

So in short, using the replica session_replication_role should be better for the only case this ever helped with and allows us to more safely use --commit=often.

As noted in the comments, this does come at the risk that the database may actually have 'replica' triggers, but it doesn't seem too inappropriate in general to fire them given their use case.