Change the default --commit from 'success' to 'often' and remove the 'tables' level
In practice, almost all users of KS re-run if they get a failure. But under --commit=success we would rollback, which under PostgreSQL would inevitably bloat the tables (sometimes quite drastically), and on MySQL/MariaDB the rollback itself could take a very long time (and re-running while this is happening generally produces lock wait timeout errors which often confuse the user who doesn't know that the rollback is still in progress).

So in reality nearly everyone is better off if they use --commit=often.

We'll describe this as a sort of breaking behavior change for the handful of people who want to get an updated database or not (and presumably expect only a small delta), though since commit isn't truly atomic anyway, arguably there isn't any previous guarantee being broken now.  In any case, they can use --commit=success to get the old default back.

Meanwhile --commit=tables was not problematic, but didn't really have any meaningful use case and didn't provide much in the way of consistency guarantees for the database as a whole, so it's not worth keeping.