Don't bother trying to commit on a failed run even with --commit=often; try to rollback explicitly in all cases
There wasn't much point trying to commit since if the --commit=often option was set we would have committed all except any very recent work, and in the failure case who knows what the situation is.

Obviously the database server would rollback anyway when the connection dies, but waiting for the rollback to complete gives a better chance of avoiding locking drama if the user re-runs.