Support for tables with no primary key and no usable unique key but that do have a key and all columns non-nullable.
This covers the 'join table' use case, where there may not be a primary key but there's a covering index and there are no extra nullable columns.

Previously this no-PK case was only supported if there was a *unique* index with all non-nullable columns.  In practice, sometimes people don't make such join table indexes unique.  Here we support them by treating the whole row as a pseudo-key for the row, and counting duplicate rows.

We could optimise it further by only deleting or inserting N-M rows in the case where we have N and should have M rows, or vice versa.  Since in practice it is very rare for there to be actual duplicate rows, and the count changes infrequently, this doesn't seem worth the extra implementation and testing effort.