Workaround the MDEV-4088 workaround, which was causing incorrect default value parsing when using the mysql version of the client library with a mariadb server (closes #83)
The root cause is that MariaDB returns version strings like "5.5.5-10.4.13-MariaDB-1:10.4.13+maria~bionic" due to a workaround made to support replication to older versions (MDEV-4088).

This then get parsed by the library and returned as version 5.5.5.  Later versions of the MariaDB client library chop off the 5.5.5- prefix, but older libraries won't and MySQL's version won't.

As a result, when KS runs using the MySQL client library to access a MariaDB server, it will incorrectly get told the server is 5.5.5 and will use the incorrect branch of schema interpretation code.

Added a test setup for this problem.