Pass arguments to the 'from' endpoint using env rather than the command line
This has been planned since v0.47 (2015), as a means of writing out the '-' hack and of making adding new options that the 'from' endpoint needs easier to do.

Pulling the trigger on it now because I'm about to add such an option.

Keep the old positional support code in the 'from' endpoint, as --via mode from any version prior to this will still send the options that way.