Write out OpenSSL as a build-dependency
We were only using it for MD5, which is deprecated (replaced by BLAKE3) and doesn't have any real performance magic in OpenSSL anyway; this implementation from Solar Designer has very similar performance on the platforms I've tested.

This means that, apart from the libstdc++ implementation, KS sources are now completely self-contained.

Many/most of the test build Dockerfiles still install libssl-dev. This is because it's required by the Ruby mysql2 gem, which is used by the unit tests.