Fix index reconciliation where unique index names sort after other key_types
This bug manifested as KitchenSync suggesting schema reconciliation
statements which could not apply because they suggested creating an
index that already existed.

This happened because when keys where discovered in both endpoints they
are sorted using the comparison operators defined on the struct. But
then their name was only considered in the schema matching. When a
unique key name sorted lexicographically after a standard key that
already existed in the 'to' end this edgecase was hit.