Fix key sorting order so the matcher algorithm works correctly
Previously the keys were sorted with unique indexes before the non-unique indexes, and then by name. But the matcher code ignored this and assumed that the keys were sorted in name order only, and would therefore incorrectly drop and recreate non-unique keys if a new unique key was received.

There's no particular reason to put the key type before the key name, as long as both are used somewhere in the sort order, so flip this to fix the issue and keep the algorithm simple.